% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/myrowcol_func.R
\name{colMin}
\alias{colMin}
\title{Form column minima}
\usage{
colMin(x, na.rm = TRUE, dims = 1L)
}
\arguments{
\item{x}{an array of two or more dimensions, containing numeric, complex, integer or logical values, or a numeric data frame.}

\item{na.rm}{logical. if \code{TRUE}, remove \code{NA}s, otherwise do not remove (and return errors or warnings)}

\item{dims}{integer.}
}
\value{
a vector of minimum values from each column
}
\description{
returns the minimum value by column for numeric arrays.
}
\details{
This function is equivalent to the use of apply with FUN = min, similar to \code{colSums()} and related functions in \code{base}.
}
\seealso{
\link{colSums}, \link{max}, \link{min}, \link{apply}
}

