% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/max_min_vectors.R
\name{min_complex}
\alias{min_complex}
\title{Minima of a vector}
\usage{
min_complex(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric, complex or character vector}

\item{na.rm}{boolean. If \code{TRUE}, any \code{NA} values will be dropped before maxima are calculated. Otherwise, returns \code{NA} if any \code{NA} values are present.}
}
\value{
a numeric, complex, or character vector of length == 1
}
\description{
Returns the minima of a vector, including vectors containing complex numbers. Essentially implements \code{min()} for complex numbers.
  Simply passes arguments through to \code{min()} if \code{x} is not of type \code{complex}.
}
\examples{
min_complex(runif(10))

vi <- c(complex(r = 1, i = 0), complex(r = 1, i = 1), complex(r = 2, i = 1), complex(r = 2, i = 2))
min_complex(vi)

vi2 <- c(complex(r = 1, i = 0), complex(r = 1, i = 1), complex(r = 2, i = 1), complex(r = 2, i = 2), complex(r = 3, i = 2), complex(r = 0, i = 1))
min_complex(vi2)
}
\seealso{
\link{min}, \link{complex}
}

